refactor: transaction logging to common.
