+++ src/app.module.ts
@@ -9,8 +9,8 @@ import { EscrowModule } from './endpoints/escrow/escrow.module';
 import { SubstrateModule } from './substrate/substrate.module';
 import { SubstrateIndexedDataModule } from './endpoints/substrate-indexed-data/substrate-indexed-data.module';
 import { EthereumIndexedDataModule } from './endpoints/ethereum-indexed-data/ethereum-indexed-data.module';
-import { TransactionLoggingModule } from './transaction-logging/transaction-logging.module';
-import { TransactionRequest } from './transaction-logging/models/transaction-request.entity';
+import { TransactionLoggingModule } from './common/transaction-logging/transaction-logging.module';
+import { TransactionRequest } from './common/transaction-logging/models/transaction-request.entity';
 import { RecaptchaModule } from './endpoints/recaptcha/recaptcha.module';
 import { CloudStorageModule } from './endpoints/cloud-storage/cloud-storage.module';
 import { BountyModule } from './endpoints/bounty/bounty.module';
