+++ src/substrate/serviceRequestEvent.ts
@@ -1,7 +1,7 @@
 import { ApiPromise } from '@polkadot/api';
 import { ethers } from 'ethers'
-import { TransactionLoggingDto } from '../transaction-logging/dto/transaction-logging.dto';
-import { TransactionLoggingService } from '../transaction-logging/transaction-logging.service';
+import { TransactionLoggingDto } from '../common/transaction-logging/dto/transaction-logging.dto';
+import { TransactionLoggingService } from '../common/transaction-logging/transaction-logging.service';
 import { MailerManager } from '../common/mailer';
 import { Logger } from '@nestjs/common';
 import { CountryService } from '../endpoints/location/country.service';
